+++ R/data.R
@@ -330,3 +330,19 @@ NULL
 
 #' @rdname R_EDSM_Subregions
 "R_EDSM_Subregions_19P3"
+
+#' SamplingSites
+#'
+#' Shapefile with regular sampling locations for most IEP surveys used between 2015 and 2021
+#'
+#' @format a sf tibble with 954 rows and 7 columns.
+#' \describe{
+#'   \item{Station}{Code used for station.}
+#'   \item{StationName}{Full name of station (if applicable).}
+#'   \item{Latitude}{Latitude in decimal degrees, WGS84.}
+#'   \item{Longitude}{Longitude in decimal degrees, WGS84.}
+#'   \item{Parameter}{Type of data collected.}
+#'   \item{Source}{Name of sampling program.}
+#'   \item{geometry}{Shapefile point coordinates.}
+#' }
+